+++ mobile/android/android-components/components/support/utils/src/main/java/mozilla/components/support/utils/DownloadUtils.kt
@@ -7,6 +7,7 @@ package mozilla.components.support.utils
 import android.net.Uri
 import android.os.Environment
 import android.webkit.MimeTypeMap
+import androidx.annotation.VisibleForTesting
 import java.io.ByteArrayOutputStream
 import java.io.File
 import java.io.UnsupportedEncodingException
@@ -54,16 +55,32 @@ object DownloadUtils {
         // Split fileName between base and extension
         // Add an extension if filename does not have one
         val extractedFileName = extractFileNameFromUrl(contentDisposition, url)
+        val sanitizedMimeType = sanitizeMimeType(mimeType)
 
         val fileName = if (extractedFileName.contains('.')) {
-            changeExtension(extractedFileName, mimeType)
+            changeExtension(extractedFileName, sanitizedMimeType)
         } else {
-            extractedFileName + createExtension(mimeType)
+            extractedFileName + createExtension(sanitizedMimeType)
         }
 
         return destinationDirectory?.let {
             uniqueFileName(Environment.getExternalStoragePublicDirectory(destinationDirectory), fileName)
         } ?: fileName
+    }
+
+    // Some site add extra information after the mimetype, for example 'application/pdf; qs=0.001'
+    // we just want to extract the mimeType and ignore the rest.
+    @VisibleForTesting
+    internal fun sanitizeMimeType(mimeType: String?): String? {
+        return if (mimeType != null) {
+            if (mimeType.contains(";")) {
+                mimeType.substringBefore(";")
+            } else {
+                mimeType
+            }
+        } else {
+            null
+        }
     }
 
     /**